Added documentation for plugins
